+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this module will be documented in this file.
 
+## [1.0.1] - 2022-07-13
+
+### Added
+
+- s3 account public access block
+- LICENSE, CONTRIBUTING, ISSUE tempaltes, examples etc
+
 ## [1.0.0] - 2022-06-28
 
 ### Added